Not Playable : Customizable
Can customize the player, Created a new player Called "Fair Nature Male Mage" Can edit the head, hair, and face. Added a new scene for customizing the character, disabled all scripts except the Player Controller and Customizer on the player.

Fix: Maybe add a save method, but enable the scripts and instantiate the prefab with the values and enable all scripts.